WIP: incomplete docs and cleanup pass
